Values and modifiers you created yourself and assigned to Data are no longer unique_ptrs internally
Instead these are currently not being cleaned up at all, that's your responsibility. Due to this
it is now possible to create your own values and modifiers on the stack or in global variables.

The two builtin modifiers are because of this now global variables.

Also fixed the issues with the modifiers where they would segfault because the parameters were in the
wrong order and were treated incorrectly. While doing this I also made some progress towards the
modify_numeric and modify_string part, this however is not functional just yet.