We can now correctly throw exceptions from expression tokens
This was needed as the incoming Variable in the ForEachStatement
has to be of type LiteralVariable, otherwise it is basically a syntax
error.

There is however one bug with this, somehow we get a segfault if we
call ParseFree in src/tokenprocessor.cpp if we're unwinding the stack.
Due to this it will currently NOT free this if we are unwinding the stack.
This is basically a memory leak..