Added a libjit exception handler
This exception handler will throw a human readable error message.
Even though this is an external library it is probably a good idea
to throw ourself out of it as the default behavior otherwise is to
exit the program, which you probably don't want. Besides it is pretty
well defined what happens in the default libjit exception handler
https://github.com/agalakhov/libjit/blob/master/jit/jit-except.c#L209