Implemented not equals for booleans and strings just like I did with equals
Also made the strcmp callback return the output of strncmp directly to make
the internals slightly less confusing. Both libjit and generated C code will
now have to compare against the output value of this call! This can however
mean that we can add something like "string1" > string2" later on and this would
work exactly like it would in C.