Got rid of the modify_string and modify_numeric stuff
It is kind of impossible to implement it this way anyway. Also the plan was
to use a Variant object as the input for the Modifiers. However we are still
internally working with Value* which we can't exactly turn into Variants
easily. We could do something like Variant(value->toString(), value->size());
but that would be kind of heavy due to the string copy. Leaving it at Value*
for now.