Made the |raw override modifier work
The output callback got one extra parameter which simply says if the Handler
should escape the value or not. On compile time, if we are a Filter (so we have
modifiers) we simply do a check if the non existing modifier "raw" is in the list.
If it is we set this flag to false, otherwise it'll be on true