Rewrote the code generator that writes the modifiers and parameters
Modifiers are now applied for just one statement instead of multiple.
This is purely needed to eventually solve the issue of not being able
to use modifiers on variables inside if statements.

Note that might require an increased -fbracket-depth in the case of
clang as the bracket nesting can go quite deep, but this entirely depends
on the amount of modifiers and parameters your statements have of course.