Implemented a simple runtime error class
Apparently we have no way of giving info about the runtime error just yet, although
that doesn't really matter as we only have one kind of runtime error which is possible
at the moment. Which is the zero division. But we might have more types of runtime errors
in the near future, which means we should obviously have some way to say something about the
occuring runtime error.