Changed an internal cast to the correct type
I'm amazed by how long this has worked without any issues though.
But yeah the value we get is a plain old Value* not a VariantValue*
however our modifer api wants a VariantValue so we could either change
this api (I for now didn't do this) to accept a Value, or like how I
solved it for now simply wrap this Value into a VariantValue.