Changed the Modifier api after all
It'll now accept a const Value& instead of a const VariantValue&
This does however mean we are unable to return the original input.
So instead of this I introduced a simple NoModification exception
which you can throw from your Modifier's to indiciate that you don't
intend to modify the value, so we know internally that we should use
the original value.